full cart backend code completed

+++ db/CartFunctions.js
@@ -6,6 +6,7 @@ const CartSchema=  mongoose.Schema({
     title:String,
     description:String,
     price:Number,
+    finalPrice:Number,
     quantity:Number,
     user : {
         _id:mongoose.Types.ObjectId
@@ -22,18 +23,53 @@ async function getAllCartProducts({page,pageLimit,sortOrder,sortBy}){
     return {totalCartProducts,cartProducts}
 }
 
+async function addToCart(userId,productData){
+  //    const user=await UserModel.findById(userId)
+  //  if(!user){
+  //     throw new Error('Please login before adding to cart userId not found')
+  //  }
+  let productInCart=await CartModel.findById(productData._id)
+     if(productInCart){
+       let status= await CartModel.updateOne({_id:productInCart._id},{
+            $set:{
+                finalPrice:productInCart.finalPrice+productInCart.price,
+                quantity:productInCart.quantity+1
+            }
+        })
+
+        let updatedProductInCart=await CartModel.findById(productData._id)
+        console.log("productInCart 3-->",updatedProductInCart)
+
+        return updatedProductInCart
+    
+     }
+     else{
+        console.log("calling inside else")
+          const productAddedInCart=await CartModel.create({
+               ...productData,
+               finalPrice:productData.price,
+               quantity:1,
+               user:{
+                   _id:userId
+               }
+           })
+           return productAddedInCart
+        
+      }
+}
+
 async function updateCartProduct(userId,productId,productData){
     let productInCart=await CartModel.findById(productId)
 
     if(!productInCart){
         throw new Error('Product not available in cart refresh or delete and add again')
     }
-    // if(String(productInCart.user._id)!==String(userId)){
-    //     throw new Error('Please login first to make changes in cart')
-    // }
-    productInCart.update({
+    if(String(productInCart.user._id)!==String(userId)){
+        throw new Error('Please login first to make changes in cart')
+    }
+    let status= await CartModel.updateOne({_id:productInCart._id},{
         $set:{
-            price:productData.price,
+            finalPrice:productData.quantity * productInCart.price,
             quantity:productData.quantity
         }
     })
@@ -42,13 +78,18 @@ async function updateCartProduct(userId,productId,productData){
 }
 
 async function deleteCartProduct(userId,productId){
+  //  const user=await UserModel.findById(userId)
+  //  if(!user){
+  //     throw new Error('Please login before adding to cart userId not found')
+  //  }
+
   let productInCart =await CartModel.findById(productId)
   if(!productInCart){
     throw new Error('Product not available in cart refresh')
   }
-  // if(String(productInCart.user._id)!==String(userId)){
- //     throw new Error('Please login first to make changes in cart')
- // }
+  if(String(productInCart.user._id)!==String(userId)){
+     throw new Error('Please login first to make changes in cart')
+ }
  productInCart=await CartModel.findByIdAndDelete(productId)
  return productInCart
 
@@ -58,5 +99,6 @@ module.exports={
     CartModel,
     getAllCartProducts,
     updateCartProduct,
-    deleteCartProduct
+    deleteCartProduct,
+    addToCart
 }
